+++ src/main/java/roomescape/domain/reservationdate/ReservationDateService.java
@@ -4,9 +4,9 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import roomescape.domain.reservation.ReservationRepository;
-import roomescape.domain.reservationdate.dto.AdminReservationDateResponse;
-import roomescape.domain.reservationdate.dto.CreateReservationDateRequest;
-import roomescape.domain.reservationdate.dto.CreateReservationDateResponse;
+import roomescape.domain.reservationdate.admin.dto.AdminReservationDateResponse;
+import roomescape.domain.reservationdate.admin.dto.CreateReservationDateRequest;
+import roomescape.domain.reservationdate.admin.dto.CreateReservationDateResponse;
 import roomescape.domain.reservationdate.dto.ReservationDateResponse;
 import roomescape.support.exception.ConflictException;
 import roomescape.support.exception.errors.ReservationDateErrors;
