+++ src/test/java/roomescape/domain/reservationdate/ReservationDateServiceTest.java
@@ -10,9 +10,9 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import roomescape.domain.reservation.Reservation;
-import roomescape.domain.reservationdate.dto.AdminReservationDateResponse;
-import roomescape.domain.reservationdate.dto.CreateReservationDateRequest;
-import roomescape.domain.reservationdate.dto.CreateReservationDateResponse;
+import roomescape.domain.reservationdate.admin.dto.AdminReservationDateResponse;
+import roomescape.domain.reservationdate.admin.dto.CreateReservationDateRequest;
+import roomescape.domain.reservationdate.admin.dto.CreateReservationDateResponse;
 import roomescape.domain.reservationtime.ReservationTime;
 import roomescape.domain.theme.Theme;
 import roomescape.support.exception.RoomescapeException;
